Profiler: use unsigned type

+++ src/platform/profiler/Profiler.cpp
@@ -169,7 +169,7 @@ public:
 		return stringIndex;
 	}
 	
-	int entries() {
+	size_t entries() {
 		return m_list.size();
 	}
